Edit project structure
fend_main.py split into run.py, models.py, routes.py, init.py

* Add models.py. Contains database documents.
* Add routes.py. Contains page routes.
* Add run.py
* Add __init__.py
* Add errors.py. Contains errorhandlers.
