+++ src/frontend/errors.py
@@ -1,0 +1,17 @@
+from src.frontend import app
+from flask import render_template
+
+
+@app.errorhandler(404)
+def page_not_found(e):
+    return render_template('/errors/page-404.html'), 404
+
+
+@app.errorhandler(403)
+def page_forbidden(e):
+    return render_template('/errors/page-403.html'), 403
+
+
+@app.errorhandler(500)
+def page_exception(e):
+    return render_template('/errors/page-500.html'), 500
