+++ src/frontend/routes.py
@@ -1,82 +1,34 @@
-from datetime import datetime
-
-from flask import Flask, render_template, request, redirect, url_for, flash
-from flask_bcrypt import check_password_hash
-from flask_login import LoginManager, current_user, login_required, login_user, UserMixin, logout_user
-from flask_mail import Mail, Message
-from flask_mongoengine import MongoEngine
+from flask import render_template, request, redirect, url_for, flash, abort
+from flask_login import current_user, login_required, login_user, logout_user
+from flask_mail import Message
 from jinja2 import TemplateNotFound
 from werkzeug.urls import url_parse
 
-app = Flask(__name__)
-
-app.config.from_object('config')
-app.config.from_object('config.MongoConfig')
-app.config.from_object('config.MailConfig')
-
-db = MongoEngine()
-db.init_app(app)
-
-login_manager = LoginManager(app)
-login_manager.login_view = 'login'
-mail = Mail(app)
-
-
-class Endpoints(db.Document):
-    date_created = db.DateTimeField(default=datetime.now())
-    url = db.StringField()
-    endpoint = db.StringField()
-
-
-class Keywords(db.Document):
-    crawl_keys = db.DictField()
-    recrawl_keys = db.DictField()
-    wanted_keys = db.DictField()
-    unwanted_keys = db.DictField()
-
-
-class User(UserMixin, db.Document):
-    meta = {'collection': 'users'}
-    username = db.StringField(unique=True)
-    password = db.StringField(default=True)
-    active = db.BooleanField(default=True)
-
-    def __repr__(self):
-        return '<User {}>'.format(self.username)
-
-    def check_password(self, password):
-        return check_password_hash(self.password, password)
-
-    def is_authenticated(self):
-        return True
-
-    def is_active(self):
-        return True
-
-    def is_anonymous(self):
-        return False
+from src.frontend import app, models, login_manager, mail
 
 
 @app.route('/', methods=['GET', 'POST'])
 def index():
     try:
-        endpoints = Endpoints.objects()
+        endpoints = models.Endpoints.objects()
         return render_template('index.html', endpoints=endpoints)
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/crawler', methods=['GET', 'POST'])
 def crawler():
     try:
-        keywords = list(Keywords.objects.exclude("id"))[0]["crawl_keys"]
+        keywords = list(models.Keywords.objects.exclude("id"))[0]["crawl_keys"]
+        if request.method == 'POST':
+            return redirect(url_for("crawler", keywords=keywords))
         return render_template('crawler.html', keywords=keywords)
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/about')
@@ -84,9 +36,9 @@ def about():
     try:
         return render_template('about.html')
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/contact', methods=['GET', 'POST'])
@@ -112,20 +64,20 @@ def contact():
             return redirect(url_for("contact"))
         return render_template('contact.html')
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/endpoint/<path:ep_url>', methods=['GET', 'POST'])
 def endpoint(ep_url):
     try:
-        endpoints = Endpoints.objects()
+        endpoints = models.Endpoints.objects()
         return render_template('endpoint.html', ep_url=ep_url, endpoints=endpoints)
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/selectedEndpoint', methods=['GET', 'POST'])
@@ -135,14 +87,14 @@ def selectedEndpoint():
             ep_url = str(list(request.form.values())[0])
             return redirect(url_for('endpoint', ep_url=ep_url))
         except TemplateNotFound:
-            return render_template('/errors/page-404.html'), 404
+            abort(404)
         except:
-            return render_template('/errors/page-500.html'), 500
+            abort(500)
 
 
 @login_manager.user_loader
 def load_user(pk):
-    return User.objects.get(pk=pk)
+    return models.User.objects.get(pk=pk)
 
 
 @app.route('/login.html', methods=('GET', 'POST'))
@@ -154,7 +106,7 @@ def login():
             form_username = request.form['username']
             form_password = request.form['userpass']
 
-            user = User.objects(username=form_username).first()
+            user = models.User.objects(username=form_username).first()
 
             if user is None or not user.check_password(form_password):
                 return redirect(url_for('login'))
@@ -166,9 +118,9 @@ def login():
 
         return render_template('/auth/login.html')
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/logout')
@@ -177,9 +129,9 @@ def logout():
         logout_user()
         return redirect(url_for('index'))
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @app.route('/dashboard', methods=['GET', 'POST'])
@@ -190,20 +142,16 @@ def dashboard():
             return render_template('index.html')
         return render_template('dashboard.html')
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
+        abort(500)
 
 
 @login_manager.unauthorized_handler
 def unauthorized_callback():
     try:
-        return render_template('/errors/page-403.html'), 403
+        return abort(403)
     except TemplateNotFound:
-        return render_template('/errors/page-404.html'), 404
+        abort(404)
     except:
-        return render_template('/errors/page-500.html'), 500
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
+        abort(500)
