+++ src/frontend/templates/crawler.html
@@ -5,7 +5,7 @@
 
 {% block body %}
 
-<form action="/crawler.html" method="POST">
+<form action="/crawler" method="POST">
     <section class="section is-large">
         <div class="container has-text-centered">
             <div class="columns is-vcentered">
